[Misc] -> Added owner to CCharacter too,
[Bugfix]
-> A summon can't attack another summon (from another player),
-> summon shouldn't attack a player or its summon in a non pvp map,
-> problem with the GM command /set (refine) solved,
-> monster shouldn't attack a player (or anything) if the target is already dead),
-> perhaps a server crash avoided in some rare cases in mainprocess,
-> skillpoints should be taken now when learning a new skill,

[Refine check]
-> little check added in charserver and worldserver in order to load the characters properly even if a refine is wrong...,
-> a warning will be issued in charserver and worldserver and the refine will be changed,
-> the changed item will be saved at the next auto save,

[Zuly]
-> (maxxon) should handle more than 2 billions now,
-> if you add 4 billions to your account using /givezuly, you'll have a negative value sent to you as a message ingame, it's a bug from client, you can see the real amount is positive,

[MySQL]
-> accounts and characters table changed because of the Zuly 2 billions stuff...,


[GM Command]
-> /summon updated:
--> /summon monsterid aipid
---> monsterid=monster id, like usual
---> aipid=Id of the AIP you want him to use (0 makes server use the monster's real AIP),

+++ Database/accounts.sql
@@ -15,7 +15,7 @@ CREATE TABLE IF NOT EXISTS accounts (
 	donation varchar(255) default '0',
 	active int(11) default 1,
 	active_key varchar(255) default NULL,
-	zulystorage int(11) default 0,
+	zulystorage bigint(11) default 0,
 	platinum tinyint(1) default 0,
 	online tinyint(1) default 0,
 	login_count int(11) default 0,
