+++ World Server/chartype.h
@@ -79,7 +79,7 @@ struct STATS
     unsigned int sidechance; //future expansion of drop system
     unsigned int ExtraDamage;
     unsigned int ExtraDamage_add;
-//    unsigned int ShieldDamage;    
+//    unsigned int ShieldDamage;
 };
 
 struct BATTLE
@@ -109,7 +109,7 @@ struct STATUS
     BYTE Dash_up;
     BYTE HP_up;
     BYTE MP_up;
-    BYTE ShieldDamage_up;    
+    BYTE ShieldDamage_up;
     //BYTE ExtraDamage_up;
     //Stats down
     BYTE Attack_down;
@@ -123,7 +123,7 @@ struct STATUS
     BYTE HP_down;
     BYTE MP_down;
     //BYTE ExtraDamage_down;
-    BYTE ShieldDamage_down;    
+    BYTE ShieldDamage_down;
 
     //LMA: ospRose skill.
     //Status Effects
@@ -232,8 +232,8 @@ struct INFO
     //long int Zulies;
     //long int Storage_Zulies;
     //DevilKing's update for more than 2 billion zuly
-    unsigned long int Zulies;
-    unsigned long int Storage_Zulies;
+    long long Zulies;
+    long long Storage_Zulies;
     long int LastGlobal;
     int StatPoints;
     int SkillPoints;
